chore: remove balance state from redux store
