--- packages/reactnative/src/store/reducers/Balance.ts
@@ -1,15 +1,0 @@
-import { createSlice } from '@reduxjs/toolkit';
-
-export const balanceSlice = createSlice({
-  name: 'BALANCE',
-  initialState: '',
-  reducers: {
-    setBalance: (state, action) => {
-      return action.payload;
-    }
-  }
-});
-
-export const { setBalance } = balanceSlice.actions;
-
-export default balanceSlice.reducer;
